requirements.txt
# Phase 2 dependencies
flask>=3.0.0         # Approval server + observability dashboard
twilio>=9.0.0        # SMS / WhatsApp / Voice alerts
boto3>=1.34.0        # AWS EC2 + CloudWatch (REAL mode only)
python-dotenv>=1.0.0 # .env support for credentials
scikit-learn>=1.4.0  # ML-based VM ranking (Isolation Forest)
numpy>=1.26.0        # Numerical operations for ML
pyyaml>=6.0.0        # YAML config support
